Configuring the administration panel to show the registered users in a dynamic way.

+++ src/main/java/com/example/vim/auth/LoginUserRequest.java
@@ -8,6 +8,6 @@ import lombok.*;
 @NoArgsConstructor
 public class LoginUserRequest {
 
-    long folio;
+    int folio;
     String email;
 }
